+++ src-docs/src/views/badge/badge_button.js
@@ -7,7 +7,8 @@ export default () => (
     <EuiBadge
       color="#333"
       onClick={() => window.alert('Badge clicked')}
-      onClickAriaLabel="Example of onclick event for the button">
+      onClickAriaLabel="Example of onclick event for the button"
+      data-test-sub="testExample1">
       onClick on badge itself
     </EuiBadge>
 
@@ -16,8 +17,21 @@ export default () => (
       iconSide="right"
       color="#333"
       iconOnClick={() => window.alert('Icon inside badge clicked')}
-      iconOnClickAriaLabel="Example of onclick event for icon within the button">
+      iconOnClickAriaLabel="Example of onclick event for icon within the button"
+      data-test-sub="testExample2">
       onClick on icon within badge
+    </EuiBadge>
+
+    <EuiBadge
+      iconType="cross"
+      iconSide="right"
+      color="#333"
+      onClick={() => window.alert('Badge clicked')}
+      onClickAriaLabel="Example of onclick event for the button"
+      iconOnClick={() => window.alert('Icon inside badge clicked')}
+      iconOnClickAriaLabel="Example of onclick event for icon within the button"
+      data-test-sub="testExample3">
+      onClick on itself and the icon
     </EuiBadge>
   </div>
 );